requirements.txt
Pillow
pillow-heif
PyYAML
